Refactor: Merge voting and staked event listeners into one common type
