+++ src/blockProcessor.ts
@@ -4,7 +4,7 @@ import {
   getMatchingListener,
 } from '~eventListeners';
 import eventProcessor from '~eventProcessor';
-import { getInterfaceByListenerType } from '~interfaces';
+import { getInterfaceByListener } from '~interfaces';
 import provider from '~provider';
 import {
   getLastBlockNumber,
@@ -49,7 +49,7 @@ export const processNextBlock = async (): Promise<void> => {
       }
 
       // In order to parse the log, we need an ether's interface
-      const iface = getInterfaceByListenerType(listener.type);
+      const iface = getInterfaceByListener(listener);
       if (!iface) {
         output(
           `Failed to get an interface for a log with listener type ${listener.type}`,
