+++ src/eventListeners/extension/index.ts
@@ -1,0 +1,96 @@
+import { utils } from 'ethers';
+
+import { ContractEventsSignatures } from '~types';
+import {
+  EventListenerType,
+  addEventListener,
+  getEventListeners,
+  setEventListeners,
+} from '~eventListeners';
+import {
+  ExtensionFragment,
+  ListExtensionsDocument,
+  ListExtensionsQuery,
+  ListExtensionsQueryVariables,
+} from '~graphql';
+import { query } from '~amplifyClient';
+import { notNull } from '~utils';
+
+import { addNetworkEventListener } from '../network';
+import { setupListenersForVotingReputationExtensions } from './votingReputation';
+import { setupListenersForStakedExpenditureExtensions } from './stakedExpenditure';
+
+export * from './stakedExpenditure';
+export * from './votingReputation';
+
+export const addExtensionEventListener = (
+  eventSignature: ContractEventsSignatures,
+  extensionHash: string,
+  extensionAddress: string,
+  colonyAddress: string,
+): void => {
+  addEventListener({
+    type: EventListenerType.Extension,
+    eventSignature,
+    address: extensionAddress,
+    colonyAddress,
+    topics: [utils.id(eventSignature)],
+    extensionHash,
+  });
+};
+
+export const removeExtensionEventListeners = (
+  extensionAddress: string,
+): void => {
+  const existingListeners = getEventListeners();
+  setEventListeners(
+    existingListeners.filter((listener) => {
+      if (listener.type !== EventListenerType.Extension) {
+        return true;
+      }
+
+      return listener.address !== extensionAddress;
+    }),
+  );
+};
+
+export const setupListenersForExtensions = async (): Promise<void> => {
+  const extensionEvents = [
+    ContractEventsSignatures.ExtensionInstalled,
+    ContractEventsSignatures.ExtensionUninstalled,
+    ContractEventsSignatures.ExtensionDeprecated,
+    ContractEventsSignatures.ExtensionUpgraded,
+  ];
+
+  extensionEvents.forEach((eventSignature) =>
+    addNetworkEventListener(eventSignature),
+  );
+
+  setupListenersForVotingReputationExtensions();
+  setupListenersForStakedExpenditureExtensions();
+};
+
+export const fetchExistingExtensions = async (
+  extensionHash: string,
+): Promise<ExtensionFragment[]> => {
+  const extensions = [];
+  let nextToken: string | undefined;
+
+  do {
+    const { data } =
+      (await query<ListExtensionsQuery, ListExtensionsQueryVariables>(
+        ListExtensionsDocument,
+        {
+          nextToken,
+          hash: extensionHash,
+        },
+      )) ?? {};
+
+    const { items } = data?.getExtensionsByHash ?? {};
+    extensions.push(...(items ?? []));
+
+    nextToken = data?.getExtensionsByHash?.nextToken ?? '';
+  } while (nextToken);
+
+  return extensions.filter(notNull);
+};
