--- src/eventListeners/extensions.ts
@@ -1,21 +1,0 @@
-import { ContractEventsSignatures } from '~types';
-
-import { addNetworkEventListener } from './network';
-import { setupListenersForVotingReputationExtensions } from './votingReputation';
-import { setupListenersForStakedExpenditureExtensions } from './stakedExpenditure';
-
-export const setupListenersForExtensions = async (): Promise<void> => {
-  const extensionEvents = [
-    ContractEventsSignatures.ExtensionInstalled,
-    ContractEventsSignatures.ExtensionUninstalled,
-    ContractEventsSignatures.ExtensionDeprecated,
-    ContractEventsSignatures.ExtensionUpgraded,
-  ];
-
-  extensionEvents.forEach((eventSignature) =>
-    addNetworkEventListener(eventSignature),
-  );
-
-  setupListenersForVotingReputationExtensions();
-  setupListenersForStakedExpenditureExtensions();
-};
