+++ src/eventListeners/index.ts
@@ -2,7 +2,5 @@ export * from './eventListeners';
 export * from './types';
 export * from './colony';
 export * from './network';
-export * from './votingReputation';
-export * from './stakedExpenditure';
-export * from './extensions';
+export * from './extension';
 export * from './token';
