+++ src/eventListeners/types.ts
@@ -10,8 +10,7 @@ export interface BaseEventListener {
 export enum EventListenerType {
   Colony = 'Colony',
   Network = 'Network',
-  VotingReputation = 'VotingReputation',
-  StakedExpenditure = 'StakedExpenditure',
+  Extension = 'Extension',
   Token = 'Token',
 }
 
@@ -25,25 +24,19 @@ interface NetworkEventListener extends BaseEventListener {
   address: string;
 }
 
-interface VotingReputationEventListener extends BaseEventListener {
-  type: EventListenerType.VotingReputation;
-  address: string;
-  colonyAddress: string;
-}
-
-interface StakedExpenditureEventListener extends BaseEventListener {
-  type: EventListenerType.StakedExpenditure;
-  address: string;
-  colonyAddress: string;
-}
-
 interface TokenEventListener extends BaseEventListener {
   type: EventListenerType.Token;
+}
+
+interface ExtensionEventListener extends BaseEventListener {
+  type: EventListenerType.Extension;
+  address: string;
+  colonyAddress: string;
+  extensionHash: string;
 }
 
 export type EventListener =
   | ColonyEventListener
   | NetworkEventListener
-  | VotingReputationEventListener
-  | StakedExpenditureEventListener
-  | TokenEventListener;
+  | TokenEventListener
+  | ExtensionEventListener;
