+++ src/handlers/extensions/extensionUninstalled.ts
@@ -1,4 +1,4 @@
-import { removeStakedExpenditureListeners } from '~eventListeners';
+import { removeExtensionEventListeners } from '~eventListeners';
 import { ContractEvent } from '~types';
 import { deleteExtensionFromEvent } from '~utils';
 
@@ -6,6 +6,5 @@ export default async (event: ContractEvent): Promise<void> => {
   const { contractAddress: extensionAddress } = event;
 
   await deleteExtensionFromEvent(event);
-  removeStakedExpenditureListeners(extensionAddress);
-  removeStakedExpenditureListeners(extensionAddress);
+  removeExtensionEventListeners(extensionAddress);
 };
