+++ src/interfaces.ts
@@ -5,18 +5,21 @@ import {
   StakedExpenditureEvents__factory as StakedExpenditureEventsFactory,
   TokenEvents__factory as TokenEventsFactory,
 } from '@colony/colony-js/events';
+import { Extension, getExtensionHash } from '@colony/colony-js';
 
 import networkClient from '~networkClient';
 import provider from '~provider';
-import { EventListenerType } from '~eventListeners';
+import { EventListener, EventListenerType } from '~eventListeners';
 
 /**
  * Function returning ABI-derived interface for a given event listener type,
  * which is later used for parsing event logs
  */
-export const getInterfaceByListenerType = (
-  listenerType: EventListenerType,
+export const getInterfaceByListener = (
+  listener: EventListener,
 ): utils.Interface | null => {
+  const { type: listenerType } = listener;
+
   switch (listenerType) {
     case EventListenerType.Network: {
       return networkClient.interface;
@@ -25,21 +28,34 @@ export const getInterfaceByListenerType = (
       return ColonyEventsFactory.connect(constants.AddressZero, provider)
         .interface;
     }
-    case EventListenerType.VotingReputation: {
+    case EventListenerType.Extension: {
+      return getInterfaceByExtensionHash(listener.extensionHash);
+    }
+    case EventListenerType.Token: {
+      return TokenEventsFactory.connect(constants.AddressZero, provider)
+        .interface;
+    }
+    default: {
+      return null;
+    }
+  }
+};
+
+const getInterfaceByExtensionHash = (
+  extensionHash: string,
+): utils.Interface | null => {
+  switch (extensionHash) {
+    case getExtensionHash(Extension.VotingReputation): {
       return VotingReputationEventsFactory.connect(
         constants.AddressZero,
         provider,
       ).interface;
     }
-    case EventListenerType.StakedExpenditure: {
+    case getExtensionHash(Extension.StakedExpenditure): {
       return StakedExpenditureEventsFactory.connect(
         constants.AddressZero,
         provider,
       ).interface;
-    }
-    case EventListenerType.Token: {
-      return TokenEventsFactory.connect(constants.AddressZero, provider)
-        .interface;
     }
     default: {
       return null;
